% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_markdown_table.R
\name{csv_to_markdown_table}
\alias{csv_to_markdown_table}
\title{Convert a csv file to a markdown table}
\usage{
csv_to_markdown_table(csv_path)
}
\arguments{
\item{csv_path}{Character. Path to csv file.}
}
\description{
Also convert two columns to markdown links
}
